Make ReactionEmoji#name raise on nil
Currently, there is a "wontfix" bug in Discord's gateway API that
causes, in rare cases, for `emoji.name` to be null for reaction events
with custom emoji. This goes against the type specification for these
payloads, where it should always be a non-nullable string. This causes
some reaction events to never reach clients, because we fail to parse them.

This patch is a minimal fix: we now can parse payloads with `null` emoji
names, but we recognize the bug by raising an exception if the client
actually needs the emoji name and it isn't there.

The proper workaround for clients is something similar as follows. The
use case shown is for tracking emoji usage.

```crystal
id = emoji.id
name = begin
  # Try to read the emoji name
  emoji.name
rescue
  # Fall back onto REST
  if id
    client.get_guild_emoji(id).name
  end
end
database.incr(name, id)
```

I think the impact of this is pretty small. In my own observations,
this only occurs specifically with `MESSAGE_REACTION_REMOVE`.
Additionally, for using the received emoji , you do not need
a custom emoji's name in order to use it with the REST API. (it is
ignored)

Both of the following examples will work for an emoji with the name
"rubytaco" and ID 315242245274075157:

```crystal
client.create_reaction(channel_id, message_id, "rubytaco:315242245274075157")
client.create_reaction(channel_id, message_id, "foo:315242245274075157")
```

Alternatively, we could:

- Make a specific `ReactionEmoji` type for `MESSAGE_REACTION_REMOVE`
payloads that has `name` as `String?`. I don't think it's right though
to make a specific type to handle a bug in the service.

- Make `ReactionEmoji#name` itself `String?`. I don't think this is
worth inconveniencing all types that use `ReactionEmoji`, when the
frequency of it occurring and number of use cases for reading the
emoji's name specifically on this event is very small.